bot: Split workflow in several hooks, fixes #14
