+++ bot/code_coverage_bot/hooks/base.py
@@ -1,0 +1,131 @@
+# -*- coding: utf-8 -*-
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+import os
+import tempfile
+
+import hglib
+import structlog
+
+from code_coverage_bot import config
+from code_coverage_bot import grcov
+from code_coverage_bot import taskcluster
+from code_coverage_bot import uploader
+from code_coverage_bot.artifacts import ArtifactsHandler
+from code_coverage_bot.utils import ThreadPoolExecutorResult
+
+logger = structlog.get_logger(__name__)
+
+
+class Hook(object):
+    def __init__(self, task_name_filter, cache_root, revision=None):
+        temp_dir = tempfile.mkdtemp()
+        self.artifacts_dir = os.path.join(temp_dir, "ccov-artifacts")
+        self.reports_dir = os.path.join(temp_dir, "ccov-reports")
+
+        assert os.path.isdir(cache_root), f"Cache root {cache_root} is not a dir."
+        self.repo_dir = os.path.join(cache_root, self.branch)
+
+        if revision is None:
+            # Retrieve latest ingested revision
+            try:
+                self.revision = uploader.gcp_latest(self.branch)[0]["revision"]
+            except Exception as e:
+                logger.warn(
+                    "Failed to retrieve the latest reports ingested: {}".format(e)
+                )
+                raise
+        else:
+            self.revision = revision
+        logger.info(
+            "Mercurial setup", repository=self.repository, revision=self.revision
+        )
+
+        task_ids = {}
+        for platform in ["linux", "windows", "android-test", "android-emulator"]:
+            task = taskcluster.get_task(self.branch, self.revision, platform)
+
+            # On try, developers might have requested to run only one platform, and we trust them.
+            # On mozilla-central, we want to assert that every platform was run (except for android platforms
+            # as they are unstable).
+            if task is not None:
+                task_ids[platform] = task
+            elif (
+                self.repository == config.MOZILLA_CENTRAL_REPOSITORY
+                and not platform.startswith("android")
+            ):
+                raise Exception("Code coverage build failed and was not indexed.")
+
+        self.artifactsHandler = ArtifactsHandler(
+            task_ids, self.artifacts_dir, task_name_filter
+        )
+
+    @property
+    def branch(self):
+        return self.repository[len(config.HG_BASE) :]
+
+    def clone_repository(self):
+        cmd = hglib.util.cmdbuilder(
+            "robustcheckout",
+            self.repository,
+            self.repo_dir,
+            purge=True,
+            sharebase="hg-shared",
+            upstream="https://hg.mozilla.org/mozilla-unified",
+            revision=self.revision,
+            networkattempts=7,
+        )
+
+        cmd.insert(0, hglib.HGPATH)
+
+        proc = hglib.util.popen(cmd)
+        out, err = proc.communicate()
+        if proc.returncode:
+            raise hglib.error.CommandError(cmd, proc.returncode, out, err)
+
+        logger.info("{} cloned".format(self.repository))
+
+    def retrieve_source_and_artifacts(self):
+        with ThreadPoolExecutorResult(max_workers=2) as executor:
+            # Thread 1 - Download coverage artifacts.
+            executor.submit(self.artifactsHandler.download_all)
+
+            # Thread 2 - Clone repository.
+            executor.submit(self.clone_repository)
+
+    def build_reports(self, only=None):
+        """
+        Build all the possible covdir reports using current artifacts
+        """
+        os.makedirs(self.reports_dir, exist_ok=True)
+
+        reports = {}
+        for (
+            (platform, suite),
+            artifacts,
+        ) in self.artifactsHandler.get_combinations().items():
+
+            if only is not None and (platform, suite) not in only:
+                continue
+
+            # Generate covdir report for that suite & platform
+            logger.info(
+                "Building covdir suite report",
+                suite=suite,
+                platform=platform,
+                artifacts=len(artifacts),
+            )
+            output = grcov.report(
+                artifacts, source_dir=self.repo_dir, out_format="covdir"
+            )
+
+            # Write output on FS
+            path = os.path.join(self.reports_dir, f"{platform}.{suite}.json")
+            with open(path, "wb") as f:
+                f.write(output)
+
+            reports[(platform, suite)] = path
+
+        return reports
